@@ -1,40 +1,37 @@
 const path = require("path");
-const HtmlWebpackPlugin= require('html-webpack-plugin');
-module.exports = {
-    entry: './src/index.js',
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin"); // npm i -D clean-webpack-plugin
 
-    output: {
-        path: path.join(__dirname,"/dist"),
-        filename: "index_bundle.js",
-    },
-    module:{
-        rules: [
-            {
-                test: /\.js$|\.jsx$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader'
-                }
-            },
-            {
-                test: /\.css$/,
-                use: [
-                    {
-                        loader: 'style-loader',
-                    },
-                    {
-                        loader: 'css-loader',
-                    },
-                ]
-            },
-        ],
-    },
-    plugins: [
-        new HtmlWebpackPlugin({
-            template: "./src/index.html"
-        })
+module.exports = {
+  entry: "./src/index.js",
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "bundle.js",
+    // clean: true      
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: "babel-loader",
+      },
+      {
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"],
+      },
     ],
-    resolve: {
-        extensions: ['.js', '.jsx'],
-    }
+  },
+  plugins: [
+    new CleanWebpackPlugin(), 
+    new HtmlWebpackPlugin({
+      template: "./public/index.html",
+    }),
+  ],
+  resolve: { extensions: [".js", ".jsx"] },
+  devServer: {
+    port: 8080,
+    open: true,
+    hot: true,
+  },
 };
